requirements.txt
apprise==1.9.3
python-dotenv==1.1.0
reportbug==12.0.0
